requirements.txt
streamlit
yt-dlp
openai-whisper
ffmpeg-python
torch